fix(lint): apply VGA 9.0.0 fixes + add gradual migration exceptions
After enabling very_good_analysis 9.0.0 in commit 3038289f, flutter analyze
revealed 3090 info issues across 30+ rule categories. This commit:

## Code fixes applied (manual)

### parameter_assignments (1 occurrence)
- lib/services/booking_derived_fields_service.dart:142
  Changed: 'if (baseRate < 0) { baseRate = 0; }' → 'final effectiveBaseRate = baseRate < 0 ? 0.0 : baseRate;'
  Reason: parameter reassignment is forbidden by VGA

### avoid_slow_async_io (2 occurrences)
- lib/services/file_management_service.dart:165,582
  Changed: 'await file.stat()' → 'file.statSync()'
  Changed: 'await entity.stat()' → 'entity.statSync()'
  Reason: file.stat() async wraps sync I/O — statSync() is faster

### avoid_bool_literals_in_conditional_expressions (1 occurrence)
- lib/screens/settings/settings_users.dart:723
  Changed: 'isAdminUser ? true : _perms.contains(k)' → 'isAdminUser || _perms.contains(k)'
  Reason: boolean literal in conditional is redundant

### avoid_catching_errors (1 occurrence)
- lib/screens/bookings/booking_edit.dart:698
  Added: '// ignore: avoid_catching_errors' with explanatory comment
  Reason: catching StateError is intentional (logical error handling)

### noop_primitive_operations (2 occurrences)
- lib/screens/payments/booking_checkout_screen.dart:172,180
  Removed: '.toDouble()' on already-double values from .clamp()
  Reason: clamp on doubles returns double — toDouble() is no-op

## Gradual migration exceptions added

Added 19 new rules to disabled list (with occurrence counts):
- no_default_cases (8) — changing default may break behavior
- prefer_int_literals (165) — doubles are legitimate (currency amounts)
- avoid_dynamic_calls (147) — JSON parsing requires dynamic
- omit_local_variable_types (457) — we prefer explicit types for clarity
- cascade_invocations (554) — cascade may reduce readability
- avoid_catches_without_on_clauses (1084) — generic catch acceptable in services
- document_ignores (270) — documenting ignore comments
- always_put_required_named_parameters_first (193) — current builder order matters
- specify_nonobvious_property_types (46)
- use_raw_strings (45)
- flutter_style_todos (22)
- avoid_returning_this (19) — Builder pattern
- unnecessary_underscores (14)
- use_setters_to_change_properties (10)
- unintended_html_in_doc_comment (10) — < > in comments
- avoid_multiple_declarations_per_line (8)
- avoid_escaping_inner_quotes (7)
- combinators_ordering (6)
- deprecated_consistency (3)

## Strategy

1. Fix low-count rules manually (done above: 6 fixes)
2. Disable high-count rules temporarily (done above: 19 exceptions)
3. Future sprints: re-enable one rule at a time after fixing all occurrences
4. Track progress: 3090 infos → 0 (after re-enabling all rules)

## Verification

Previous run (3038289f): flutter analyze = success (with 3090 infos)
Expected this run: flutter analyze = success (with 0 infos after exceptions)

+++ mobile/lib/screens/payments/booking_checkout_screen.dart
@@ -169,7 +169,7 @@ class _BookingCheckoutScreenState extends ConsumerState<BookingCheckoutScreen>
                           final discountedRate = (roomPrice - discount).clamp(
                             0.0,
                             roomPrice,
-                          ).toDouble();
+                          );
                           return (fullNights * roomPrice) +
                               (discountedNights * discountedRate);
                         }
@@ -177,7 +177,7 @@ class _BookingCheckoutScreenState extends ConsumerState<BookingCheckoutScreen>
                       })();
 
                 final totalDue = discount > 0 && discountType == 'total'
-                    ? (nightTotal - discount).clamp(0.0, nightTotal).toDouble()
+                    ? (nightTotal - discount).clamp(0.0, nightTotal)
                     : nightTotal;
 
                 return Padding(
